# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/bank_accounts:
    get:
      tags:
        - BankAccounts
      summary: List BankAccounts
      description: Get a paginated list of BankAccounts with optional filtering and sorting
      operationId: listBankAccounts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BankAccounts
      summary: Create BankAccount
      description: Create a new BankAccount
      operationId: createBankAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankAccountRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankAccount'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/bank_clearances:
    get:
      tags:
        - BankClearances
      summary: List BankClearances
      description: Get a paginated list of BankClearances with optional filtering and sorting
      operationId: listBankClearances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankClearanceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BankClearances
      summary: Create BankClearance
      description: Create a new BankClearance
      operationId: createBankClearance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBankClearanceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankClearance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    BankAccount:
      type: object
      required:
        - id
        - company_id
        - bank_id
        - account_name
        - account_number
        - gl_account_id
        - clearing_account_id
        - currency
        - account_type
        - is_default
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        bank_id:
          type: string
          format: uuid
        account_name:
          type: string
          maxLength: 140
        account_number:
          type: string
          maxLength: 60
        gl_account_id:
          type: string
          format: uuid
        clearing_account_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        account_type:
          type: string
        is_default:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateBankAccountRequest:
      type: object
      required:
        - company_id
        - bank_id
        - account_name
        - account_number
        - gl_account_id
        - clearing_account_id
        - currency
        - account_type
        - is_default
        - is_active
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        bank_id:
          type: string
          format: uuid
        account_name:
          type: string
          maxLength: 140
        account_number:
          type: string
          maxLength: 60
        gl_account_id:
          type: string
          format: uuid
        clearing_account_id:
          type: string
          format: uuid
        currency:
          type: string
          maxLength: 3
        account_type:
          type: string
        is_default:
          type: boolean
        is_active:
          type: boolean
        metadata:
          type: object
...
    BankAccountList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BankAccount'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    BankClearance:
      type: object
      required:
        - id
        - company_id
        - bank_transaction_id
        - matched_source_type
        - matched_source_id
        - matched_amount
        - match_method
        - clearance_date
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        bank_transaction_id:
          type: string
          format: uuid
        matched_source_type:
          type: string
        matched_source_id:
          type: string
          format: uuid
        matched_amount:
          type: string
        match_method:
          type: string
        clearance_date:
          type: string
          format: date
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        journal_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreateBankClearanceRequest:
      type: object
      required:
        - company_id
        - bank_transaction_id
        - matched_source_type
        - matched_source_id
        - matched_amount
        - match_method
        - clearance_date
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        bank_transaction_id:
          type: string
          format: uuid
        matched_source_type:
          type: string
        matched_source_id:
          type: string
          format: uuid
        matched_amount:
          type: string
        match_method:
          type: string
        clearance_date:
          type: string
          format: date
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        journal_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    BankClearanceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BankClearance'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []