# src/presentation/http/bank_account_handler.rs
//! BankAccount REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{BankAccountService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateBankAccountDto, UpdateBankAccountDto, PatchBankAccountDto, BankAccountResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum BankAccountError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for BankAccountError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for BankAccountError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "BANKACCOUNT_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "BANKACCOUNT_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKACCOUNT_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKACCOUNT_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for BankAccount.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /bank_accounts | List with pagination |
/// | POST | /bank_accounts | Create new |
/// | GET | /bank_accounts/:id | Get by ID |
/// | PUT | /bank_accounts/:id | Full update |
/// | PATCH | /bank_accounts/:id | Partial update |
/// | DELETE | /bank_accounts/:id | Soft delete |
/// | POST | /bank_accounts/bulk | Bulk create |
/// | POST | /bank_accounts/upsert | Upsert |
/// | GET | /bank_accounts/trash | List deleted |
/// | POST | /bank_accounts/:id/restore | Restore |
/// | DELETE | /bank_accounts/empty | Empty trash |
/// | GET | /bank_accounts/:id/deleted | Get deleted by ID |
/// | DELETE | /bank_accounts/trash/:id | Permanent delete from trash |
/// | GET | /bank_accounts/count | Count active entities |
/// | GET | /bank_accounts/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(BankAccountService::with_repository(repository));
/// let router = create_bank_account_routes(service);
/// ```
pub fn create_bank_account_routes(service: Arc<BankAccountService>) -> Router {
    BackboneCrudHandler::<BankAccountService, BankAccount, CreateBankAccountDto, UpdateBankAccountDto, BankAccountResponseDto>::routes(
        service,
        "/bank_accounts",
    )
}

/// Create Axum router with only the read (GET) endpoints for BankAccount.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_bank_account_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_bank_account_read_routes(service: Arc<BankAccountService>) -> Router {
    BackboneCrudHandler::<BankAccountService, BankAccount, CreateBankAccountDto, UpdateBankAccountDto, BankAccountResponseDto>::read_routes(
        service,
        "/bank_accounts",
    )
}

/// Create Axum router with only the write (mutation) endpoints for BankAccount.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_bank_account_write_routes(service: Arc<BankAccountService>) -> Router {
    BackboneCrudHandler::<BankAccountService, BankAccount, CreateBankAccountDto, UpdateBankAccountDto, BankAccountResponseDto>::write_routes(
        service,
        "/bank_accounts",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_bank_account_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<BankAccountService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_bank_account_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}