# src/presentation/http/bank_clearance_handler.rs
//! BankClearance REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{BankClearanceService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateBankClearanceDto, UpdateBankClearanceDto, PatchBankClearanceDto, BankClearanceResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum BankClearanceError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for BankClearanceError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for BankClearanceError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "BANKCLEARANCE_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "BANKCLEARANCE_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKCLEARANCE_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "BANKCLEARANCE_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for BankClearance.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /bank_clearances | List with pagination |
/// | POST | /bank_clearances | Create new |
/// | GET | /bank_clearances/:id | Get by ID |
/// | PUT | /bank_clearances/:id | Full update |
/// | PATCH | /bank_clearances/:id | Partial update |
/// | DELETE | /bank_clearances/:id | Soft delete |
/// | POST | /bank_clearances/bulk | Bulk create |
/// | POST | /bank_clearances/upsert | Upsert |
/// | GET | /bank_clearances/trash | List deleted |
/// | POST | /bank_clearances/:id/restore | Restore |
/// | DELETE | /bank_clearances/empty | Empty trash |
/// | GET | /bank_clearances/:id/deleted | Get deleted by ID |
/// | DELETE | /bank_clearances/trash/:id | Permanent delete from trash |
/// | GET | /bank_clearances/count | Count active entities |
/// | GET | /bank_clearances/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(BankClearanceService::with_repository(repository));
/// let router = create_bank_clearance_routes(service);
/// ```
pub fn create_bank_clearance_routes(service: Arc<BankClearanceService>) -> Router {
    BackboneCrudHandler::<BankClearanceService, BankClearance, CreateBankClearanceDto, UpdateBankClearanceDto, BankClearanceResponseDto>::routes(
        service,
        "/bank_clearances",
    )
}

/// Create Axum router with only the read (GET) endpoints for BankClearance.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_bank_clearance_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_bank_clearance_read_routes(service: Arc<BankClearanceService>) -> Router {
    BackboneCrudHandler::<BankClearanceService, BankClearance, CreateBankClearanceDto, UpdateBankClearanceDto, BankClearanceResponseDto>::read_routes(
        service,
        "/bank_clearances",
    )
}

/// Create Axum router with only the write (mutation) endpoints for BankClearance.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_bank_clearance_write_routes(service: Arc<BankClearanceService>) -> Router {
    BackboneCrudHandler::<BankClearanceService, BankClearance, CreateBankClearanceDto, UpdateBankClearanceDto, BankClearanceResponseDto>::write_routes(
        service,
        "/bank_clearances",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_bank_clearance_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<BankClearanceService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_bank_clearance_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}